Move NDData I/O functions to astropy.io.fits.connect
