+++ astropy/io/fits/connect_table.py
@@ -3,13 +3,11 @@
 from __future__ import print_function
 
 import os
-import re
 import warnings
 
 import numpy as np
 
 from .. import registry as io_registry
-from ... import log
 from ... import units as u
 from ...extern import six
 from ...extern.six import string_types
@@ -23,62 +21,7 @@ from . import FITS_rec
 from .hdu.hdulist import fitsopen as fits_open
 from .util import first
 
-
-# FITS file signature as per RFC 4047
-FITS_SIGNATURE = (b"\x53\x49\x4d\x50\x4c\x45\x20\x20\x3d\x20\x20\x20\x20\x20"
-                  b"\x20\x20\x20\x20\x20\x20\x20\x20\x20\x20\x20\x20\x20\x20"
-                  b"\x20\x54")
-
-# Keywords to remove for all tables that are read in
-REMOVE_KEYWORDS = ['XTENSION', 'BITPIX', 'NAXIS', 'NAXIS1', 'NAXIS2',
-                   'PCOUNT', 'GCOUNT', 'TFIELDS']
-
-# Column-specific keywords
-COLUMN_KEYWORDS = ['TFORM[0-9]+',
-                   'TBCOL[0-9]+',
-                   'TSCAL[0-9]+',
-                   'TZERO[0-9]+',
-                   'TNULL[0-9]+',
-                   'TTYPE[0-9]+',
-                   'TUNIT[0-9]+',
-                   'TDISP[0-9]+',
-                   'TDIM[0-9]+',
-                   'THEAP']
-
-
-def is_column_keyword(keyword):
-    for c in COLUMN_KEYWORDS:
-        if re.match(c, keyword) is not None:
-            return True
-    return False
-
-
-def is_fits(origin, filepath, fileobj, *args, **kwargs):
-    """
-    Determine whether `origin` is a FITS file.
-
-    Parameters
-    ----------
-    origin : str or readable file-like object
-        Path or file object containing a potential FITS file.
-
-    Returns
-    -------
-    is_fits : bool
-        Returns `True` if the given file is a FITS file.
-    """
-    if fileobj is not None:
-        pos = fileobj.tell()
-        sig = fileobj.read(30)
-        fileobj.seek(pos)
-        return sig == FITS_SIGNATURE
-    elif filepath is not None:
-        if filepath.lower().endswith(('.fits', '.fits.gz', '.fit', '.fit.gz')):
-            return True
-    elif isinstance(args[0], (HDUList, TableHDU, BinTableHDU, GroupsHDU)):
-        return True
-    else:
-        return False
+from .connect_common import _put_meta_in_header, _put_header_in_meta, is_fits
 
 
 def read_table_fits(input, hdu=None):
@@ -167,29 +110,7 @@ def read_table_fits(input, hdu=None):
 
     # TODO: deal properly with unsigned integers
 
-    for key, value, comment in table.header.cards:
-
-        if key in ['COMMENT', 'HISTORY']:
-            if key in t.meta:
-                t.meta[key].append(value)
-            else:
-                t.meta[key] = [value]
-
-        elif key in t.meta:  # key is duplicate
-
-            if isinstance(t.meta[key], list):
-                t.meta[key].append(value)
-            else:
-                t.meta[key] = [t.meta[key], value]
-
-        elif (is_column_keyword(key.upper()) or
-              key.upper() in REMOVE_KEYWORDS):
-
-            pass
-
-        else:
-
-            t.meta[key] = value
+    _put_header_in_meta(table.header, t.meta)
 
     # TODO: implement masking
 
@@ -264,31 +185,7 @@ def write_table_fits(input, output, overwrite=False):
                     "The unit '{0}' could not be saved to FITS format".format(
                         unit.to_string()), AstropyUserWarning)
 
-    for key, value in input.meta.items():
-
-        if is_column_keyword(key.upper()) or key.upper() in REMOVE_KEYWORDS:
-
-            warnings.warn(
-                "Meta-data keyword {0} will be ignored since it conflicts "
-                "with a FITS reserved keyword".format(key), AstropyUserWarning)
-
-        if isinstance(value, list):
-            for item in value:
-                try:
-                    table_hdu.header.append((key, item))
-                except ValueError:
-                    warnings.warn(
-                        "Attribute `{0}` of type {1} cannot be written to "
-                        "FITS files - skipping".format(key, type(value)),
-                        AstropyUserWarning)
-        else:
-            try:
-                table_hdu.header[key] = value
-            except ValueError:
-                warnings.warn(
-                    "Attribute `{0}` of type {1} cannot be written to FITS "
-                    "files - skipping".format(key, type(value)),
-                    AstropyUserWarning)
+    _put_meta_in_header(input.meta, table_hdu.header)
 
     # Write out file
     table_hdu.writeto(output)
