+++ astropy/io/fits/tests/test_connect.py
@@ -1,5 +1,7 @@
+from __future__ import (absolute_import, division, print_function,
+                        unicode_literals)
+
 import os
-import sys
 import warnings
 
 import numpy as np
@@ -8,10 +10,11 @@ from numpy.testing import assert_allclose
 from ... import fits
 from .. import HDUList, PrimaryHDU, BinTableHDU
 from ....table import Table
+from ....nddata import NDData, NDIOMixin
 from .... import units as u
-from .... import log
+from ....extern import six
 from ....tests.helper import pytest, catch_warnings
-from astropy.units.format.fits import UnitScaleError
+from ....units.format.fits import UnitScaleError
 
 DATA = os.path.join(os.path.dirname(__file__), 'data')
 
@@ -269,3 +272,36 @@ def test_bool_column(tmpdir):
     with fits.open(str(tmpdir.join('test.fits'))) as hdul:
         assert hdul[1].data['col0'].dtype == np.dtype('bool')
         assert np.all(hdul[1].data['col0'] == arr)
+
+
+# Tests for the NDData class
+
+
+def test_simple_write_read(tmpdir):
+    path = tmpdir.join('tmp.fits')
+    data_in = np.array([1, 2, 3])
+    ndd = NDData(data_in)
+    ndd.write(path.strpath)
+    # Does the data read directly by io.fits match?
+    with fits.open(path.strpath) as f:
+        np.testing.assert_array_equal(data_in, f[0].data)
+    # Does the data read by our reader match?
+    ndd2 = NDData.read(path.strpath)
+    np.testing.assert_array_equal(data_in, ndd2.data)
+
+
+def test_write_read_with_meta(tmpdir):
+    path = tmpdir.join('tmp.fits')
+    data_in = np.array([1, 2, 3])
+    meta_in = {"one": 1, "two": 2}
+    ndd = NDData(data_in, meta=meta_in)
+    ndd.write(path.strpath)
+    # Does the data read directly by io.fits contain the meta?
+    # [Note that it will also contain other metadata.]
+    with fits.open(path.strpath) as f:
+        for k, v in six.iteritems(meta_in):
+            print(k, v)
+            print(f[0].header)
+            assert f[0].header[k] == v
+
+    ndd2 = NDData.read(path.strpath)
